% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{calcPars}
\alias{calcPars}
\title{Function calcPars}
\usage{
calcPars(x, ctr)
}
\arguments{
\item{x}{A matrix with two features on which to calculate the correlations (etc)}

\item{ctr}{The center to which each observation belong}
}
\value{
A list of a list with the following components
  \item{cc}{Correlations value per center}
  \item{N}{Number of non-NA entries in the dataset per center}
  \item{sd}{Marginal sds per center (one per line)}
}
\description{
Calculate correlations, margin s.d. on each center in a dataset.
}
\seealso{
\code{\link{testCorFisher}}, \code{\link{testCorMargin}}
}
