% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSMtests-package.R
\docType{data}
\name{centerSizes}
\alias{centerSizes}
\title{Distribution of center sizes in real studies.}
\format{
## `centerSizes`
A list of length 3 (small, medium, large). 
Each item of the list is the N patients of a center from that study
}
\usage{
centerSizes
}
\description{
Distribution of center sizes (N patients) for 3 real studies of different sizes.
}
\keyword{datasets}
