% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSMtests-package.R
\docType{package}
\name{CSMtests}
\alias{CSMtests}
\title{CSMtests}
\description{
Tools for central statistical monitoring.
}
\details{
At the moment nothing but correlation-based tests.
}
\examples{
Np = round(runif(100)*50+10);
x = simulateCentersCors(.5, .1, Np)
ctr = rep(1:100, Np);

pp = calcPars(x, ctr)

p1 = testCorFisher(pp$cc, pp$N)
p2 = testCorMargin(pp$cc, pp$N, pp$sd)
}
\seealso{
\code{\link{testCorFisher}}, \code{\link{testCorMargin}}m \code{\link{calcPars}}
}
