% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{simulateCentersCors}
\alias{simulateCentersCors}
\title{Function simulateCentersCors}
\usage{
simulateCentersCors(cc, rho, Np)
}
\arguments{
\item{cc}{The mean correlation}

\item{rho}{Standard deviation on the correlation}

\item{Np}{Number of patients for each center. Number of centers drawn is given by the length of Np.}
}
\value{
A matrix with 2 columns with the bivariate normal data
}
\description{
Simulate a set of centers drawn from a bivariate distribution, with a given average
correlation and a given standard deviation on that correlation (on the Fisher scale).
}
