% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.r
\name{testCorFisher}
\alias{testCorFisher}
\title{Function testCorFisher}
\usage{
testCorFisher(corrs, Ni, retAll = FALSE)
}
\arguments{
\item{corrs}{A vector with the correlations per center}

\item{Ni}{Number of data points used to calculate the correlations, per center}

\item{retAll}{if FALSE, return only the p-values per center}
}
\value{
If retAll is TRUE, A list of a list with the following components
  \item{p}{P-values per center. P-values signs indicate whether centers are higher or lower than expected. }
  \item{mu}{Mean correlation}
  \item{raneff}{Standard deviation on the correlation)}
}
\description{
Apply the Fisher test on a set of correlations.
}
\seealso{
\code{\link{calcPars}}, \code{\link{testCorMargin}}
}
